extract one more cached strategy to its own file
